Refactor PantryItem and PantryAdd into IngredientItem and IngredientAdd components to use across app for all edit/add ingredient processes
