+++ client/src/actions/pantryActions.js
@@ -1,7 +1,6 @@
 import axios from 'axios';
 import isEmpty from 'is-empty';
 import cookies from 'js-cookie';
-import {REQUEST_FAIL} from './types';
 
 // Add ingredient to pantry
 export const addIngredientToPantry = async (ingredientData) => {
@@ -33,19 +32,16 @@ export const addIngredientToPantry = async (ingredientData) => {
 
   if (!isEmpty(error)) {
     return {
-      authResponseType: REQUEST_FAIL,
-      authResponsePayload: error,
+      data: error,
     };
   } else {
     try {
       const token = cookies.get('user');
-      const response = await axios.post('/api/pantry', ingredientData, {
+      return await axios.post('/api/pantry', ingredientData, {
         headers: {
           Authorization: token,
         },
       });
-
-      return response;
     } catch (err) {
       return err.response.data;
     }
@@ -75,13 +71,12 @@ export const updateIngredientInPantry = async (ingredientData) => {
 
   if (!isEmpty(error)) {
     return {
-      authResponseType: REQUEST_FAIL,
-      authResponsePayload: error,
+      data: error,
     };
   } else {
     try {
       const token = cookies.get('user');
-      const response = await axios.put(
+      return await axios.put(
         `/api/pantry/${ingredientData.id}`,
         ingredientData,
         {
@@ -90,8 +85,6 @@ export const updateIngredientInPantry = async (ingredientData) => {
           },
         }
       );
-
-      return response;
     } catch (err) {
       return err.response.data;
     }
@@ -112,19 +105,16 @@ export const deleteIngredientFromPantry = async (ingredientId) => {
 
   if (!isEmpty(error)) {
     return {
-      authResponseType: REQUEST_FAIL,
-      authResponsePayload: error,
+      data: error,
     };
   } else {
     try {
       const token = cookies.get('user');
-      const response = await axios.delete(`/api/pantry/${ingredientId}`, {
+      return await axios.delete(`/api/pantry/${ingredientId}`, {
         headers: {
           Authorization: token,
         },
       });
-
-      return response;
     } catch (err) {
       return err.response.data;
     }
